Fix the CI jobs that went red during the encoding ports
MSVC has no max_align_t in C mode, so strings/replace.c uses
BURROW_ALIGN_MAX like the rest of the tree. The amalgamation failed on
a parameter named run shadowing a static in strconv. clang-tidy
findings in encoding, strings and bytes are fixed, and binary.c
initialises its targets so gcc at -O3 stops warning.

FuzzTypes failed on s390x and under msan because its two seeds run in
parallel and shared one static buffer for the caught panic text. Each
caller has its own buffer now.

+++ src/bytes/bytes.c
@@ -713,7 +713,7 @@ Slice bytes_repeat(Alloc *a, Slice b, Int count) {
     /* Past a certain chunk size it is counterproductive to use larger chunks
      * as the source of the write, as when the source is too large we are
      * basically just thrashing the CPU D-cache. */
-    const Int chunk_limit = 8 * 1024;
+    const Int chunk_limit = (Int)8 * 1024;
     Int chunk_max = n;
     if (chunk_max > chunk_limit) {
         chunk_max = chunk_limit / b.len * b.len;
